Graphics.UI.Bucephalus.Core

Documentation

class (CoreInterface a b, GamePad p) => GameState a b s p | s -> p where

This is type class for holding game status.

Methods

gameMainCore :: a -> (p, s) -> IO s

This function is main programm of your game.

gameQuitCore :: a -> s -> IO ()

At the time of an end, bucephalus core program would call this function.

coreStart :: (CoreInterface a b, GamePad p, GameState a b s p) => CoreConf a -> p -> s -> IO ()

coreStart function provide basic operations. e.g. : main loop operation ,end judging and so on.

data StandardPad

StandardPad data type provide game pad type, that has basic buttons.

Constructors

StandardPad

Fields

buttonUp :: Integer

buttonDown :: Integer

buttonLeft :: Integer

buttonRight :: Integer

buttonA :: Integer

buttonB :: Integer

buttonX :: Integer

buttonY :: Integer

buttonL :: Integer

buttonR :: Integer

buttonStart :: Integer

buttonSelect :: Integer

padInit :: GamePad p => p

This function would return initial game pad.

padToVector :: StandardPad -> (Int, Int)

This function is utility for StandardPad. e.g. : When pushing buttons UP and RIGHT, this function would return (-1, 1).

data CoreConf a

This is configuration data for bucephalus Core program.

Constructors

CoreConf

Fields

fullScreen :: Bool

coreInterface :: a

class CoreInterface a b | a -> b where

This type class provide interface to multi media libraly.

Methods

bucephalusInit :: CoreConf a -> IO ()

bucephalusQuit :: CoreConf a -> IO ()

bucephalusGetTicks :: a -> IO Word32

bucephalusPollEvent :: a -> IO BucephalusEvent

bucephalusDelay :: a -> Word32 -> IO ()

bucephalusLoadImg :: a -> String -> IO b

bucephalusBulitImg :: a -> b -> Maybe (Int, Int, Int, Int) -> b -> Maybe (Int, Int, Int, Int) -> IO ()

bucephalusFreeImg :: a -> b -> IO ()

bucephalusRotoZoom :: a -> b -> Double -> Double -> Bool -> IO b

bucephalusGetWidth :: a -> b -> Int

bucephalusGetHeight :: a -> b -> Int

bucephalusGetVideoFrame :: a -> IO b

bucephalusFlip :: a -> b -> IO ()

bucephalusFreeFrame :: a -> b -> IO ()

bucephalusFillRect :: a -> b -> Maybe (Int, Int, Int, Int) -> Word32 -> IO ()